scripts/requirements.txt
google-analytics-data==0.18.16
google-api-python-client==2.151.0
google-auth==2.35.0
